requirements.txt
Flask==2.0.3
mysql-connector-python==8.0.28
google-api-python-client==2.39.0
